enable sample() caching in LMTP
